Updated skeleton file with changes from test file
Greatly enhanced the multiple enemy systems. Now enemies function independently so that if an enemy dies, the others continue to live. They have their own vision and range to attack the player so that if one enemy is attacking the player, an enemy across the map will continue to stand still.