Updated maps and added kills concept
Added a system so that multiple maps can transition when the player reaches the end of one map. Added a condition so that player must kill every enemy in current stage to switch to next stage, with a gui that pops up if the player tries to proceed without fulfilling that condition.